refactor: extract session service tests
